# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation for Dihauti Residence",
        "description": "description",
        "contact": {
            "email": "dev.arven@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.dihauti.ru/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/questions/{forumQuestion}/replies/{id}/like": {
            "post": {
                "tags": [
                    "Like"
                ],
                "summary": "Like reply for question",
                "operationId": "replyLike",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Question slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reply for question post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Liked this post successfully"
                    },
                    "202": {
                        "description": "Unliked this post successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
        "/contacts": {
            "get": {
                "tags": [
                    "Messenger"
                ],
                "summary": "Get contact list",
                "description": "Returns contact list",
                "operationId": "getContactList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/messages": {
            "get": {
                "tags": [
                    "Messenger"
                ],
                "summary": "Get all current dialog",
                "description": "Returns all current dialog",
                "operationId": "getCurrentDialog",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/messages/{dialog}": {
            "get": {
                "tags": [
                    "Messenger"
                ],
                "summary": "Get messages in dialog",
                "description": "Get messages in dialog",
                "operationId": "getMessagesInDialog",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Dialog slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/dialog/{dialog}": {
            "get": {
                "tags": [
                    "Messenger"
                ],
                "summary": "Get dialog",
                "description": "Get dialog",
                "operationId": "getDialog",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Dialog slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/messages/{dialog}/send": {
            "post": {
                "tags": [
                    "Messenger"
                ],
                "summary": "Create message in database",
                "operationId": "sendNewMessage",
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
        "/startDialog": {
            "post": {
                "tags": [
                    "Messenger"
                ],
                "summary": "Create gialog in database",
                "operationId": "startDialog",
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/login",
                        "refreshUrl": "/logout",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    }
}